% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposition-package.R
\docType{package}
\name{transposition-package}
\alias{transposition}
\alias{transposition-package}
\title{transposition: Transposition}
\description{
Détermine si une matrice est une matrice de transposition ou pas
}
\author{
\strong{Maintainer}: NKOUNGHAWE Rosalie Corine \email{corine.nkounghawe@facsciences-uy1.cm} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
