% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transposition.R
\name{transposition}
\alias{transposition}
\title{Transposition}
\usage{
transposition(M, i, j)
}
\arguments{
\item{M}{Any Matrix}

\item{i}{The row index of the non-zero value of Eij: May help to know which rows to swap}

\item{j}{The column index of the non-zero value of Eij: Can help to know the columns to permute}
}
\value{
Boolean: TRUE if the matrix is a transpose matrix and FALSE otherwise
}
\description{
Determines whether a matrix is a transpose matrix or not. This information can help developers to do new experiments on matrices
}
\examples{
transposition(array(c(1,0,0,0,0,1,0,1,0), dim = c(3,3)), 2, 3): This returns TRUE
transposition(array(c(2,5,0,0,2,1,0,1,2), dim = c(3,3)), 2, 3): this return FALSE
}
